verification_service/requirements.txt
flask
flask-cors
pymongo
face-recognition
opencv-python
numpy
Pillow
python-dotenv
requests
setuptools

